\name{lineRatio-package}
\alias{lineRatio-package}
\alias{lineRatio}
\docType{package}
\title{
Fragment Length and GC bias Analysis
}
\description{
Provides method use to perform Fragment-length and GC bias analysis. This package is customized to the requirements of MalariaGEN in the analysis P.falciparum data.
}
\details{
\tabular{ll}{
Package: \tab FLGCA\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-09-20\cr
License: \tab GPLv3\cr
Depends: \tab R (>= 2.14.1), mclust (>= 4.0), IRanges (>= 1.12.6), ggplot2 (>= 0.9.1), MASS (>= 7.3), rtracklayer (>= 1.14.4), bitops (>= 1.0), Biostrings (>= 2.22), lineRatio (>= 0.1)\cr
}
}
\author{Valentin Ruano-Rubio <valentin.ruano@gmail.com>}
\references{}
\keyword{ DNA fragment-length GC bias }
\seealso{
}
\examples{
}
